renamed react component folder, and abstracted the logic to their own js files
