+++ components/Header.js
@@ -1,0 +1,14 @@
+import React from "react";
+import "./App.css";
+export default function Header() {
+  return (
+    <>
+      <nav className="header">
+        <h1>Impartial News</h1>
+        <h2>Home</h2>
+        <h2>Explore</h2>
+        <h2>Lists</h2>
+      </nav>
+    </>
+  );
+}
